feat: impl parallel walk dir(but this is slow)
