+++ cspell.jsonc
@@ -6,6 +6,7 @@
     "Geoffroy",
     "Greatsword",
     "icns",
+    "jwalk",
     "multispace",
     "rfind",
     "Rustup",
