+++ dar2oar_core/Cargo.toml
@@ -17,8 +17,15 @@ nom = { version = "7.1.3", features = ["alloc"] }         # Syntax
 serde = { version = "1.0", features = ["derive"] }        # Implement (De)Serializer
 serde_json = "1.0"                                        # Json converter
 thiserror = "1.0.48"                                      # define errors type
-walkdir = "2.4.0"                                         # To traverse dir recursivly
+jwalk = "0.8.1"                                           # To parallel traverse dir recursivly
+walkdir = "2.4.0"                                         # To traverse dir recursivly(Single thiread)
 
 [dev-dependencies]
+criterion = { version = "0.5.1", features = ["html_reports"] }
 pretty_assertions = "1.4.0"
 simple-log = "1.6.0"
+
+
+[[bench]]
+name = "convert_n_thread"
+harness = false
