+++ dar2oar_core/src/lib.rs
@@ -1,7 +1,7 @@
 mod condition_parser;
 mod conditions;
 mod dar_syntax;
-mod fs;
+pub mod fs;
 mod values;
 
 pub use crate::fs::{convert_dar_to_oar, read_mapping_table};
